20.12.2025 refactor: Info section front end side

+++ qr_menu_app/view/create_social_media_icon_function.py
@@ -1,12 +1,18 @@
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponseForbidden
 from django.shortcuts import get_object_or_404, redirect, render
 from django.views.decorators.http import require_POST
+
 from qr_menu_app.forms import NavbarSocialMediaForm
 from qr_menu_app.models.navbar_model import MainSection, NavbarSocialMedia
 
 
 @login_required
+@require_POST
 def add_social_media(request, navbar_id):
+    if not (request.user.is_staff or request.user.is_superuser):
+        return HttpResponseForbidden(status=403)
+
     navbar = get_object_or_404(MainSection, id=navbar_id)
 
     if request.method == "POST":
@@ -25,6 +31,9 @@ def add_social_media(request, navbar_id):
 @login_required
 @require_POST
 def delete_social_media(request, pk):
+    if not (request.user.is_staff or request.user.is_superuser):
+        return HttpResponseForbidden(status=403)
+
     obj = get_object_or_404(NavbarSocialMedia, pk=pk)
     obj.delete()
     return redirect("home")
